+++ src/Farmer/Builders/Builders.RouteServer.fs
@@ -69,7 +69,7 @@ type RouteServerConfig =
                     Location = location
                     Sku = PublicIpAddress.Sku.Standard
                     AllocationMethod = PublicIpAddress.AllocationMethod.Static
-                    AddressVersion = PublicIpAddress.AddressVersion.IPv4
+                    AddressVersion = Network.AddressVersion.IPv4
                     DomainNameLabel = None
                     Tags = this.Tags
                 }
@@ -77,7 +77,7 @@ type RouteServerConfig =
                 //subnet
                 {
                     Subnet.Name = ResourceName "RouteServerSubnet"
-                    Prefix = IPAddressCidr.format this.SubnetPrefix
+                    Prefixes = [ IPAddressCidr.format this.SubnetPrefix ]
                     VirtualNetwork = Some(vnetId)
                     NetworkSecurityGroup = None
                     Delegations = []